requirements.txt
python-dotenv
langchain
langgraph
langchain-google-genai
langchain-mcp-adapters
llm-prompt-shield[langchain]
mcp
fastapi
python-keycloak
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlite3
opentelemetry-instrumentation-logging